+++ app/src/main/res/layout/fragment_lyrics.xml
@@ -1,0 +1,32 @@
+<?xml version="1.0" encoding="utf-8"?>
+
+<ScrollView xmlns:android="http://schemas.android.com/apk/res/android"
+    android:id="@+id/lyrics"
+    android:layout_width="match_parent"
+    android:layout_height="match_parent">
+
+    <LinearLayout
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:orientation="vertical">
+
+        <TextView
+            android:id="@+id/lyrics_text"
+            android:layout_width="wrap_content"
+            android:layout_height="wrap_content"
+            android:layout_margin="20dp"
+            android:textColor="#ededed"
+            android:textSize="30sp" />
+
+        <TextView
+            android:id="@+id/lyrics_makeitpersonal"
+            android:layout_width="wrap_content"
+            android:layout_height="wrap_content"
+            android:layout_margin="20dp"
+            android:autoLink="web"
+            android:linksClickable="true"
+            android:text="@string/lyrics_makeitpersonal"
+            android:textColor="#ededed"
+            android:textSize="17sp" />
+    </LinearLayout>
+</ScrollView>
